Fix timehelper conversions

+++ src/Time/TimeHelper.php
@@ -11,7 +11,7 @@ class TimeHelper
 
     public static function minutes(int|float $minutes): int
     {
-        return $minutes * 60 * 1000_000_000;
+        return (int) ($minutes * 60 * 1000_000_000);
     }
 
     public static function second(): int
@@ -21,7 +21,7 @@ class TimeHelper
 
     public static function seconds(int|float $seconds): int
     {
-        return $seconds * 1000_000_000;
+        return (int) ($seconds * 1000_000_000);
     }
 
     public static function millisecond(): int
@@ -31,7 +31,7 @@ class TimeHelper
 
     public static function milliseconds(int|float $milliseconds): int
     {
-        return $milliseconds * 1000_000;
+        return (int) ($milliseconds * 1000_000);
     }
 
     public static function microsecond(): int
@@ -41,7 +41,7 @@ class TimeHelper
 
     public static function microseconds(int|float $microseconds): int
     {
-        return $microseconds * 1_000;
+        return (int) ($microseconds * 1_000);
     }
 
     public static function phpMicroTime(int|float $microtime): int
